requirements.txt
ordered-set==3.1.1
Pillow==9.0.0
pyaml==21.10.1
PyLaTeX==1.4.1
PyYAML==6.0
